refactor: isolate package parsing
